+++ server/src/lib/database.ts
@@ -50,6 +50,17 @@ async function updateLine(sql: string) {
 async function updateDb() {
   import("./insert");
 
+  updateLine(`
+  CREATE TABLE \`call_types\` (
+    \`id\` varchar(255) NOT NULL,
+    \`name\` varchar(255) NOT NULL,
+    \`defaults\` varchar(255) NOT NULL DEFAULT '0',
+    PRIMARY KEY (\`id\`)
+  ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
+  
+  `);
+
+  updateLine("ALTER TABLE `911calls` ADD `type` varchar(255) NOT NULL AFTER `status`;");
   updateLine("ALTER TABLE `10_codes` ADD `position` int(255) NOT NULL AFTER `should_do`;");
   updateLine("ALTER TABLE `officers` DROP `started_at`;");
   updateLine(`